Created basic application with camera controls
There's a lot of placeholder stuff here, such as app initialisation
and control bindings, but this is all in service of getting the main
camera class set up and testable.

The camera is configured by two main parameters: its target (what it's
pointing at) and its focus (the polar coordinates defining from where
it points and at what angle/skew). This allows us to select a new
target without changing the camera angle, and allows us to rotate the
camera around a target, or bring it closer or further from a target
while allowing the camera class to worry about all the complicated
maths. I have animated these movements using a linear interpolation
set to start and end slow in order to give a 'bouncy' feel to the
camera controls and reinforce the fact that we are snapping to a grid
position, rather than having total camera freedom. In addition to this
I have locked the distance and latitude ranges the camera may move
within to keep the camera strictly above its target, and between 5 and
30 world units. These numbers may be revised later, but from hands-on
testing, these 'feel' right for a tactical game