Add basic grid-map with models and lighting
The bulk of the work here is to import models created in blender and
simplify them to a format compatible with the procedural generation I
wish to add later. I have chosen to export the geometry as json for
ease of inspection and manual tweaking, but I may add more to the json
format in use later if there are extra features I shall require.

Currently I have a class as a namespace for my colour palette, but
this is something I ay choose to move into the model json format later
depending on how the complexity of a global colour palette works out.

In order to test out the showing and hiding of geometry I have made a
basic implementation of cell selection - I intend to change this later
to allow mouse-selection, but the added callbacks will be needed for
this too.

At some point I intend to add a bloom shader so the emission source on
the grid highlight is more obvious. As the outline is made up of flat
planes, this obviously needs to be two-sided, however I should
re-order the vertices of the triangles that make up the highlight so
that these can be single-sided to make rendering cheaper. Currently I'm
trying to convince myself that this is unnecessary as there will only
ever be one of this model on display at any point, and it's fairly
low-poly with no normal mapping, but I am as of yet unsure whether
this will affect later work such as bloom shading.