Added a Bloom filter for 80's neon aesthetic
I've based the bloom filter on the lighting alpha chanel which didn't
seem to previously do anything in order to have striking bloom on
emission sources, but have reflections from other sources be more muted

There was rather a lot going on in the main init so I've subclassed
ShowNasw to do more of its own init. While doing this it made sense to
rename references to refer to it as 'base' rather than 'app' as this
seems to be the internal standard in all the documentation.